Modified how to re-enable old SS_ConfigStaticManifest to suit the SS config system
